requirements.txt
Flask==2.3.3
Werkzeug==2.3.6
Jinja2==3.1.2
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.0
matplotlib==3.8.1